Fix Carbon:createFromTimestamp data type
[MAILPOET-2590]

+++ tests/integration/Subscription/PagesTest.php
@@ -46,8 +46,8 @@ class PagesTest extends \MailPoetTest {
     $subscription->confirm();
     $confirmed_subscriber = Subscriber::findOne($this->subscriber->id);
     expect($confirmed_subscriber->status)->equals(Subscriber::STATUS_SUBSCRIBED);
-    expect($confirmed_subscriber->last_subscribed_at)->greaterOrEquals(Carbon::createFromTimestamp(current_time('timestamp'))->subSecond(1));
-    expect($confirmed_subscriber->last_subscribed_at)->lessOrEquals(Carbon::createFromTimestamp(current_time('timestamp'))->addSecond(1));
+    expect($confirmed_subscriber->last_subscribed_at)->greaterOrEquals(Carbon::createFromTimestamp((int)current_time('timestamp'))->subSecond(1));
+    expect($confirmed_subscriber->last_subscribed_at)->lessOrEquals(Carbon::createFromTimestamp((int)current_time('timestamp'))->addSecond(1));
   }
 
   public function testItDoesNotConfirmSubscriptionOnDuplicateAttempt() {
